[TASK] Update functional tests
Remove tests that depend on line numbers as they are
very unflexible and the complete files are compared anyway.

Use current (TYPO3 13.2-dev) schemas for tests,
drop fluidviewhelper tests.

In the future tests can be updated by regenerating the schema files
in the functional tests fixtures - or by exchanging them and regenerating.